database module -version catalog implemented -add language and category entity, and corresponding repository

+++ database-module/src/main/kotlin/com/news360horizon/news360horizon/database/BaseEntity.kt
@@ -6,7 +6,7 @@ import java.sql.Timestamp
 import java.time.Instant
 
 abstract class BaseEntity {
-        abstract val id: BigInteger
+        abstract val id: Long
         @set:Column(name = "is_active")
         abstract var isActive: Boolean
